// mc/src/store/modules/watch/mutations-types.ts
export enum MutationsTypes {
  VIDEO_REF = 'VIDEO_REF',
  VIDEO_STATE = 'VIDEO_STATE',
  VIDEO_SONG = 'VIDEO_SONG',
  VIDEO_LIST = 'VIDEO_LIST',
  VIDEO_INDEX = 'VIDEO_INDEX',
  VIDEO_ORDER = 'VIDEO_ORDER',
  VIDEO_VOLUME = 'VIDEO_VOLUME',
  VIDEO_FOLD = 'VIDEO_FOLD',
  VIDEO_CURRENTTIME = 'VIDEO_CURRENTTIME',
  VIDEO_DURATION = 'VIDEO_DURATION',
  VIDEO_LYRIC = 'VIDEO_LYRIC',
}